Restructure the project by adding the user types
