+++ src/Routes.jsx
@@ -1,9 +1,9 @@
-import LogIn from "./app/login/index.jsx";
-import NotFoundPage from "./app/not-found/index.jsx";
-import HomeLabManager from "./app/HomeLabManager/index.jsx";
-import SampleAssignation from "./app/SampleAssignation/index.jsx";
-import SampleManagement from "./app/SampleManagement/index.jsx";
-import Layout from "./components/layout";
+import LogIn from "./app/LabManager/login";
+import NotFoundPage from "./app/LabManager/not-found";
+import HomeLabManager from "./app/LabManager/HomeLabManager";
+import Sample from "./app/LabManager/samples";
+import ResultsAudit from "./app/LabManager/resultsAudit";
+import Layout from "./components/LabManager/layout";
 
 export const Routes = [
   { path: "*", element: <NotFoundPage /> },
@@ -13,8 +13,8 @@ export const Routes = [
     element: <Layout />,
     children: [
       { path: "homeLabManager", element: <HomeLabManager /> },
-      { path: "sampleAssignation", element: <SampleAssignation /> },
-      { path: "sampleManagement", element: <SampleManagement /> },
+      { path: "samples", element: <Sample /> },
+      { path: "resultsAudit", element: <ResultsAudit /> },
     ],
   },
 ];
