+++ src/app/LabManager/samples/index.jsx
@@ -1,0 +1,9 @@
+// import "../../styles/SampleManagement/SampleManagement.css";
+
+export default function Sample() {
+  return (
+    <div>
+      <h1>Samples</h1>
+    </div>
+  );
+}
