+++ src/app/resultsAudit/index.jsx
@@ -1,9 +1,9 @@
 // import "../../styles/SampleManagement/SampleManagement.css";
 
-export default function SampleAssignation() {
+export default function ResultsAudit() {
   return (
     <div>
-      <h1>SampleManagement</h1>
+      <h1>Results and Audits</h1>
     </div>
   );
 }
