+++ src/styles/LabManager/layout/index.css
@@ -19,9 +19,11 @@
 
 #content {
   display: flex;
-  color: blue;
+  background-color: white;
   max-width: 100vw;
   max-height: 100vh;
   box-sizing: border-box;
-  padding: 40px;
+  padding: 18px;
+  margin: 1rem;
+  border-radius: 5px;
 }
